GUA-72: Redirect `/account/home` to Digital Identity
We are moving all central account features off GOV.UK and into a DI
owned app to better separate out concerns between DI (the ID provider)
and `www.gov.uk` (a relying party).

The main page that needs to move is this one. We want to show users
which services they've used on their account home page, which isn't
possible if it doesn't sit within DI where it can access that usage
data.

Redirect the route that this app serves the account home on to a DI
domain. This is done at the app level (rather than publishing a
redirect content item with Router) so that it can do environment-aware
redirects using `GovukPersonalisation`.

Note: this implementation keeps the controller with a single `show`
method which does the redirect. I think it would be more 'Railsy' to
do the redirect in `routes.rb`, but when I tried that the application
kept trying to follow the redirect and render the path from the
destination URL (`/manage-your-account`) as a content item when tested.

Keeping the controller seemed a more pragmatic solution compared to
messing with the 'catch-all' routing fallbacks.